🎨 design: 별 아이콘 추가
