+++ src/views/library/components/verticalview-items/bookitem-items/RecordBtn.tsx
@@ -2,7 +2,11 @@ import download from '../../../../../assets/button/library/download.png';
 import file2 from '../../../../../assets/button/library/file2.png';
 import { useTabStore } from '../../../../../store/library/useTabStore';
 
-const RecordBtn = () => {
+type RecordBtnProps = {
+  text: string;
+};
+
+const RecordBtn = ({ text }: RecordBtnProps) => {
   const selectedTab = useTabStore((state) => state.selectedTab);
   return (
     <div className="flex justify-center items-center gap-4">
@@ -19,7 +23,7 @@ const RecordBtn = () => {
           className="w-9 h-9 max-w-[18px] max-h-[18px] object-contain"
         />
       )}
-      <span className="text-[1.6rem] font-normal pt-0.1">내 기록</span>
+      <span className="text-[1.6rem] font-normal pt-0.1">{text}</span>
     </div>
   );
 };
