+++ src/views/library/components/verticalview-items/bookitem-items/StarMaker.tsx
@@ -1,4 +1,5 @@
-import { FaStar } from 'react-icons/fa';
+import starFill from '../../../../../assets/button/library/star-fill.png';
+import starSubtract from '../../../../../assets/button/library/Subtract.png';
 
 interface StarMakerProps {
   star: number;
@@ -6,15 +7,20 @@ interface StarMakerProps {
 
 const StarMaker = ({ star }: StarMakerProps) => {
   const stars = [];
-  for (let i = 0; i < star; i++) {
+
+  for (let i = 0; i < 5; i++) {
     stars.push(
       <span key={i}>
-        <FaStar size={16} color="gold" />
+        <img
+          src={i < star ? starFill : starSubtract}
+          alt="별"
+          className="w-4 h-4 max-w-[13px] max-h-[13px] object-contain"
+        />
       </span>
     );
   }
 
-  return <div className="flex items-center">{stars}</div>;
+  return <div className="flex items-center gap-1">{stars}</div>;
 };
 
 export default StarMaker;
